added Decimal field product_routers: - get_product - create_product - update_product - delete_product added - pre-commit-config.yaml - methods: get_all_item, get_item to AbstractDAO category_routers added: - get_categories - get_category - update_category - create_category - delete_category

+++ models/product_models.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from pydantic import AnyUrl, BaseModel
 
 
@@ -8,7 +10,7 @@ class CategoryModel(BaseModel):
 class ProductModel(BaseModel):
     name_product: str
     picture_link: AnyUrl
-    price: float
+    price: Decimal
     product_detail_link: AnyUrl
     characteristic: dict
     description: str
